+++ python_ta/__init__.py
@@ -31,7 +31,7 @@ except AttributeError:
 import logging
 import tokenize
 import webbrowser
-from typing import IO, Any, Optional, Union
+from typing import IO, Any, Literal, Optional, Union
 
 from .check.helpers import (
     check_file,
@@ -63,6 +63,7 @@ def check_errors(
     output: Optional[Union[str, IO]] = None,
     load_default_config: bool = True,
     autoformat: Optional[bool] = False,
+    on_verify_fail: Literal["log", "raise"] = "log",
 ) -> PythonTaReporter:
     """Check a module for errors, printing a report."""
     return _check(
@@ -72,6 +73,7 @@ def check_errors(
         output=output,
         load_default_config=load_default_config,
         autoformat=autoformat,
+        on_verify_fail=on_verify_fail,
     )
 
 
@@ -81,6 +83,7 @@ def check_all(
     output: Optional[Union[str, IO]] = None,
     load_default_config: bool = True,
     autoformat: Optional[bool] = False,
+    on_verify_fail: Literal["log", "raise"] = "log",
 ) -> PythonTaReporter:
     """Analyse one or more Python modules for code issues and display the results.
 
@@ -106,6 +109,10 @@ def check_all(
             If False, the default PythonTA configuration is not used.
         autoformat:
             If True, autoformat all modules using the black formatting tool before analyzing code.
+        on_verify_fail:
+            Determines how to handle files that cannot be checked. If set to "log" (default), an error
+            message is logged and execution continues. If set to "raise", an error is raised immediately to stop
+            execution.
 
     Returns:
         The ``PythonTaReporter`` object that generated the report.
@@ -117,6 +124,7 @@ def check_all(
         output=output,
         load_default_config=load_default_config,
         autoformat=autoformat,
+        on_verify_fail=on_verify_fail,
     )
 
 
@@ -127,6 +135,7 @@ def _check(
     output: Optional[Union[str, IO]] = None,
     load_default_config: bool = True,
     autoformat: Optional[bool] = False,
+    on_verify_fail: Literal["log", "raise"] = "log",
 ) -> PythonTaReporter:
     """Check a module for problems, printing a report.
 
@@ -141,6 +150,8 @@ def _check(
     `load_default_config` is used to specify whether to load the default .pylintrc file that comes
     with PythonTA. It will load it by default.
     `autoformat` is used to specify whether the black formatting tool is run. It is not run by default.
+    `on_verify_fail` determines how to handle files that cannot be checked. If set to "log" (default), an error
+     message is logged and execution continues. If set to "raise", an error is raised immediately to stop execution.
     """
     # Configuring logger
     logging.basicConfig(format="[%(levelname)s] %(message)s", level=logging.INFO)
@@ -154,7 +165,10 @@ def _check(
             f_paths = []
             for file_py in get_file_paths(locations):
                 linted_files.add(file_py)
-                if not verify_pre_check(file_py, linter.config.allow_pylint_comments):
+                if not verify_pre_check(
+                    file_py, linter.config.allow_pylint_comments, on_verify_fail=on_verify_fail
+                ):
+                    # The only way to reach this is if verify_pre_check returns False, and `on_verify_fail="log"`.
                     continue
                 is_any_file_checked, linter = check_file(
                     file_py=file_py,